Fix #50 by zooming in a level
The reportFocus() function handles the zoom since that seems to be where NVDA first tells MathCAT it should navigate.
An alternative would be in `script_navigate`. In there, `gesture` is `None`. It seems cleaner in reportFocus().

I needed to remove the 'SetMathML' call in `__init__` because that wipes out the navigation state.
As far as I could see, nothing bad happens with that call removed.