Hack fix to #51 by monkey-patching the eSpeak speak() method.
I'm not totally sure why the fix works. The fix is to save the rate before the underlying speak() call and reset the rate after the call. However, the bug involves `<break>` and the speech isn't finished (probably not even started) by the time the reset happens. But it does work...

There is no other change to the eSpeak method, so hopefully there is no interference with other calls to it.

Note: because it was easiest, the monkey-patching is done on a espeak instance, not on the class itself. This is also a little safer because it won't happen unless someone is reading math.