Updating pyproject.toml and .pre-commit-config.yaml (#2)
This PR includes the following changes:

The files .pre-commit-config.yaml and pyproject.toml were added. The structure of these files is kept close to that found in the corresponding files from the main NVDA repository, with some configurations removed that are not applicable to this repository.
The JSON configuration file for pyright was removed, since pyright configuration is now being handled by pyproject.toml.
The option reportMissingImports was temporarily set to false, since it was triggering pyright errors.
Some unnecessary noqa comments were removed, and one was added to suppress a pyright error caused by the scriptHandler.script decorator.
Whitespace was removed from the ends of lines, CRLF was converted to LF, and other minor formatting changes were carried out by running ruff-format.
Note: To run the pre-commit checks, install pyright, ruff, and pre-commit in a virtual environment first.